# perf: introduce AST to avoid reparse