--- a/src/builtin/tags/comment.ts
+++ b/src/builtin/tags/comment.ts
@@ -1,16 +1,16 @@
-import { TagToken } from '../../parser/tag-token'
-import { Token } from '../../parser/token'
+import { TagToken } from '../../tokens/tag-token'
+import { TopLevelToken } from '../../tokens/toplevel-token'
 import { TagImplOptions } from '../../template/tag/tag-impl-options'
 
 export default {
-  parse: function (tagToken: TagToken, remainTokens: Token[]) {
+  parse: function (tagToken: TagToken, remainTokens: TopLevelToken[]) {
     const stream = this.liquid.parser.parseStream(remainTokens)
     stream
       .on('token', (token: TagToken) => {
         if (token.name === 'endcomment') stream.stop()
       })
       .on('end', () => {
-        throw new Error(`tag ${tagToken.raw} not closed`)
+        throw new Error(`tag ${tagToken.getText()} not closed`)
       })
     stream.start()
   }