--- a/src/builtin/tags/increment.ts
+++ b/src/builtin/tags/increment.ts
@@ -1,13 +1,10 @@
-import { assert } from '../../util/assert'
-import { identifier } from '../../parser/lexical'
 import { isNumber, stringify } from '../../util/underscore'
-import { Emitter, TagToken, Context, TagImplOptions } from '../../types'
+import { Tokenizer, Emitter, TagToken, Context, TagImplOptions } from '../../types'
 
 export default {
   parse: function (token: TagToken) {
-    const match = token.args.match(identifier)
-    assert(match, `illegal identifier ${token.args}`)
-    this.variable = match![0]
+    const tokenizer = new Tokenizer(token.args)
+    this.variable = tokenizer.readWord().content
   },
   render: function (context: Context, emitter: Emitter) {
     const scope = context.environments