--- a/src/builtin/tags/layout.ts
+++ b/src/builtin/tags/layout.ts
@@ -1,29 +1,25 @@
-import { assert } from '../../util/assert'
-import { quotedLine, quoted } from '../../parser/lexical'
-import { Emitter, Hash, Expression, TagToken, Token, Context, TagImplOptions } from '../../types'
+import { assert, evalQuotedToken, TypeGuards, evalToken, Tokenizer, Emitter, Hash, TagToken, TopLevelToken, Context, TagImplOptions } from '../../types'
 import BlockMode from '../../context/block-mode'
 
-const rFile = new RegExp(`^(${quoted.source}|[^\\s,]+)`)
-
 export default {
-  parse: function (token: TagToken, remainTokens: Token[]) {
-    const match = rFile.exec(token.args)
-    if (!match) {
-      throw new Error(`illegal argument "${token.args}"`)
-    }
-    this.file = match[1]
-    this.hash = new Hash(token.args.slice(match[0].length))
+  parse: function (token: TagToken, remainTokens: TopLevelToken[]) {
+    const tokenizer = new Tokenizer(token.args)
+    const file = this.liquid.options.dynamicPartials ? tokenizer.readValue() : tokenizer.readFileName()
+    assert(file, () => `illegal argument "${token.args}"`)
+
+    this.file = file
+    this.hash = new Hash(tokenizer.remaining())
     this.tpls = this.liquid.parser.parse(remainTokens)
   },
   render: function * (ctx: Context, emitter: Emitter) {
     const { liquid, hash, file } = this
     const { renderer } = liquid
     const filepath = ctx.opts.dynamicPartials
-      ? (quotedLine.exec(file)
-        ? yield renderer.renderTemplates(liquid.parse(file.slice(1, -1)), ctx)
-        : yield new Expression(file).value(ctx))
-      : this.file
-    assert(filepath, `illegal filename "${file}":"${filepath}"`)
+      ? (TypeGuards.isQuotedToken(file)
+        ? yield renderer.renderTemplates(liquid.parse(evalQuotedToken(file)), ctx)
+        : evalToken(this.file, ctx))
+      : file.getText()
+    assert(filepath, () => `illegal filename "${file.getText()}":"${filepath}"`)
 
     // render the remaining tokens immediately
     ctx.setRegister('blockMode', BlockMode.STORE)