--- a/src/parser/filter-arg.ts
+++ b/src/parser/filter-arg.ts
@@ -1,9 +1,10 @@
 import { isArray } from '../util/underscore'
+import { ValueToken } from '../tokens/value-token'
 
-type KeyValuePair = [string?, string?]
+type KeyValuePair = [string?, ValueToken?]
 
-export type FilterArg = string|KeyValuePair
+export type FilterArg = ValueToken | KeyValuePair
 
-export function isKeyValuePair (arr: FilterArg): arr is KeyValuePair {
+export function isKeyValuePair (arr: FilterArg): arr is KeyValuePair { // TODO check
   return isArray(arr)
 }