@harttle 2020-03-14
:tada: This PR is included in version 9.11.3 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/liquidjs)
- [GitHub release](https://github.com/harttle/liquidjs/releases/tag/v9.11.3)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: